requirements.txt
openai
Pillow
python-dotenv
Requests
streamlit
Markdown
PyMuPDF
chromadb
protobuf==3.20.*
pysqlite3-binary
